requirements.txt.txt
Flask
flask-cors
flask-jwt-extended
flask-limiter
python-dotenv
openai
pytesseract
python-docx
PyMuPDF
Pillow
